Refactor leader election implementation: move files to dedicated directory and update references
