+++ core/api-server/lib/examples/pipelines-updater.js
@@ -6,7 +6,7 @@ const pipelines = require('./pipelines.json');
 const drivers = require('./drivers.json');
 const experiments = require('./experiments.json');
 const stateManager = require('../state/state-manager');
-const leaderElection = require('../state/leader-election');
+const leaderElection = require('../leader-election/leader-election');
 const algorithmsVersionsService = require('../../lib/service/algorithm-versions');
 const pipelinesVersionsService = require('../../lib/service/pipeline-versions');
 const keycloak = require('../../lib/service/keycloak');
