+++ core/api-server/lib/service/internal.js
@@ -1,6 +1,6 @@
 const { pipelineTypes } = require('@hkube/consts');
 const stateManager = require('../state/state-manager');
-const leaderElection = require('../state/leader-election');
+const leaderElection = require('../leader-election/leader-election');
 const validator = require('../validation/api-validator');
 const execution = require('./execution');
